% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asks.R
\name{ask_do}
\alias{ask_do}
\title{Ask first and run}
\usage{
ask_do(what, FUN, yes = FALSE)
}
\arguments{
\item{what}{the text description of what \code{FUN} is for}

\item{FUN}{the script to be excuted}

\item{yes}{for running in a script. If 'yes' is TRUE, then 'FUN' will be 
excuted without asking}
}
\value{
if 'y' was given for the question, then output of \code{FUN}.
  Otherwise, \code{NULL} will be returned.
}
\description{
Ask first and run the \code{FUN} to avoid any unwanted automatic action in 
a script.
}
\author{
Mun-Gwan Hong, \email{mungwan@gmail.com}
}
