% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asks.R
\name{ask_save}
\alias{ask_save}
\title{Ask first and save}
\usage{
ask_save(..., list = character(), file = stop("'file' must be specified"),
  yes = FALSE)
}
\arguments{
\item{..., list, file}{Please refer to the description in \code{\link{save}}}

\item{yes}{for running in a script. If 'yes' is TRUE, then 'FUN' will be 
excuted without asking}
}
\description{
It works just like \code{save}, but ask first to avoid any unwanted automatic 
action in a script.
}
\seealso{
\code{\link{save}}
}
\author{
Mun-Gwan Hong, \email{mun-gwan.hong@scilifelab.se}
}
