% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ask_save_plot.R
\name{ask_save_plot}
\alias{ask_save_plot}
\alias{save_plot}
\title{Ask to save or to show plot}
\usage{
ask_save_plot(filename = "Rplot.png", ..., FUN)

save_plot(filename = "Rplot.png", ..., FUN)
}
\arguments{
\item{filename}{the name of file to save}

\item{...}{arguments to be passed over to a picture generation function such
as \code{\link{png}}, \code{\link{pdf}}. The function is determined by the 
the extension of \code{filename}. Currently "png", "pdf", "tif", and "jpg"
file format can be used.}

\item{FUN}{the function that make one or more plots}
}
\description{
Ask interactively if the user wants to save plot into a file or show it on
screen
}
\section{Functions}{
\itemize{
\item \code{save_plot}: Same as \code{ask_save_plot} but skip the asking. 
This is a code for swift switch between \code{ask_save_plot} and 
\code{save_plot}.
}}

\seealso{
\code{\link{png}}, \code{\link{pdf}}, \code{\link{tiff}}, 
         \code{\link{jpeg}}
}
\author{
Mun-Gwan Hong, \email{mun-gwan.hong@scilifelab.se}
}
