% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asks.R
\name{ask_write.table}
\alias{ask_write.table}
\title{Ask first and write a table}
\usage{
ask_write.table(
  x,
  file = stop("'file' must be specified"),
  quote = FALSE,
  sep = "\\t",
  na = "",
  row.names = FALSE,
  col.names = if (row.names == TRUE) NA else TRUE,
  ...,
  sortBy,
  yes = FALSE
)
}
\arguments{
\item{x, file, quote, sep, na, row.names, col.names, ...}{Please refer to the 
description in \code{\link{write.table}}}

\item{sortBy}{the column by which the rows will be sorted.}

\item{yes}{for running in a script. If 'yes' is TRUE, then 'FUN' will be 
excuted without asking}
}
\description{
It works just like \code{write.table}, but ask first to avoid any unwanted
automatic action in a script.
}
\seealso{
\code{\link{write.table}}
}
\author{
Mun-Gwan Hong, \email{mungwan@gmail.com}
}
