% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxplot_beeswarm.R
\name{boxplot_beeswarm}
\alias{boxplot_beeswarm}
\title{Box plot and Beeswarm}
\usage{
boxplot_beeswarm(formula, log = "", ...)
}
\arguments{
\item{formula, log}{same as those for \code{\link{boxplot}} and
\code{\link{boxplot_nEle}}}

\item{...}{arguments to be passed to \code{\link{boxplot_nEle}} and
eventually to \code{\link{boxplot}}}
}
\description{
Box plot and individual data points added by 'beeswarm'
}
\details{
This display a box plot together with the number of elements of each group
and individual data points drawn by 'beeswarm'
}
\seealso{
\code{\link{boxplot}}, \code{\link{boxplot_nEle}}
}
\author{
Mun-Gwan Hong, \email{mun-gwan.hong@scilifelab.se}
}
