% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxplot_nEle.R
\name{boxplot_nEle}
\alias{boxplot_nEle}
\title{Box plot and number}
\usage{
boxplot_nEle(formula, data, ..., subset, na.action, test = c("kruskal.test",
  "aov", "t.test"), test_txt_par = list(), nEle_txt_par = list())
}
\arguments{
\item{formula, data, subset, na.action}{same as those for \code{\link{boxplot}}}

\item{...}{arguments to be passed to \code{\link{boxplot}}}

\item{test}{the test to check the difference between groups. If this is 
\code{NULL} or "", no test will be performed.}

\item{test_txt_par}{a list of parameters to control on displaying the test 
result. refer to \code{\link{text}}. In here, \code{x} can be 
\code{"center"}, \code{"right"}, or numerical vectors as \code{x} of 
\code{\link{text}}. The default is \code{x = "center", y= par("usr")[4],
labels = 'pval', adj = c(0.5, 1.5), col = "navy", cex = 0.8)}.}

\item{nEle_txt_par}{a parameter list for displaying the number of elements of
each group. The default is \code{list(x = seq_along(nEle), y =
par("usr")[3], labels = nEle, col = "brown", cex = 0.8, pos = 3)}.}
}
\description{
Make a box-and-whisker plot with showing the number of elements in each group
If Y-axis is in log-scale, outliers are determined after log-transformation, 
unlike \code{boxplot}.
}
\examples{
data <- data.frame("y"= rnorm(20), "x"= gl(2, 10))
boxplot_nEle(y ~ x, data, test= "t.test", test_txt_par= list("col"= "blue"))

}
\seealso{
\code{\link{boxplot}}
}
\author{
Mun-Gwan Hong, \email{mun-gwan.hong@scilifelab.se}
}
