% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string_utils.R
\name{concatenate_skip_overlap}
\alias{concatenate_skip_overlap}
\title{Concatenate with skip}
\usage{
concatenate_skip_overlap(x, sep = ",", na.rm = TRUE)
}
\arguments{
\item{x}{a string vector}

\item{sep}{separating character}

\item{na.rm}{if NA is removed}
}
\value{
a string
}
\description{
Concatenate string with "," skipping the overlapped part at the beginning
e.g. "Sample1", "Sample2", & "Sample13" -> "Sample1,2,13"
}
\examples{
concatenate_skip_overlap(c("Sample1", "Sample2", "Sample13"))

}
\author{
Mun-Gwan Hong, \email{mun-gwan.hong@scilifelab.se}
}
