% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/english_cap1.R
\name{english_cap1}
\alias{english_cap1}
\alias{english}
\title{English and Capitalize 1}
\usage{
english_cap1(x, Zero = "No")

english(x)
}
\arguments{
\item{x}{an integer vector}

\item{Zero}{how to read 0}
}
\value{
a text
}
\description{
Read number in English and capitalize the first letter
\code{english_} fix the problem of showing digits in Rmd.
}
\examples{
english_cap1(c(1, 2))
english(c(1, 2))
}
