% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enum_vars.R
\name{enum_vars}
\alias{enum_vars}
\title{Enumerate}
\usage{
enum_vars(x, surround = "`", with_oxford_comma = FALSE)
}
\arguments{
\item{x}{a vector}

\item{surround}{with which character each element will be surrounded}

\item{with_oxford_comma}{whether an oxford comma should be added or not}
}
\value{
a text
}
\description{
Enumerate `x` in text
}
\examples{
enum_vars(c(LETTERS[1:3]))
}
