% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_model_thru.R
\name{fit_model_thru}
\alias{fit_model_thru}
\title{Fit a model through}
\usage{
fit_model_thru(dep_var_list, data, rhs, fitmodel, ...)
}
\arguments{
\item{dep_var_list}{the set of variables to be used as the dependent variable}

\item{data}{a data frame}

\item{rhs}{right hand side of the formula used in the \code{fitmodel}}

\item{fitmodel}{a function for fitting a model.}

\item{...}{this will be delivered to the \code{fitmodel} function}
}
\value{
a tibble (or data frame)
}
\description{
Fit a model changing the dependent variable through a given set of variables
and collect results
}
