% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_split_violin.R
\name{geom_split_violin}
\alias{geom_split_violin}
\title{Split violin plot}
\usage{
geom_split_violin(
  mapping = NULL,
  data = NULL,
  stat = "ydensity",
  position = "identity",
  ...,
  draw_quantiles = NULL,
  trim = TRUE,
  scale = "area",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE
)
}
\arguments{
\item{mapping, data, stat, position, ..., draw_quantiles, trim, scale, na.rm, show.legend, inherit.aes}{check \code{\link{geom_violin}}}

\item{file}{a Excel file from the manager}
}
\description{
Violin plot showing different distribution on the left and right sides. The
code of 'jan-glx' was obtained from Stackoverflow (check the reference).
}
\references{
\url{https://stackoverflow.com/questions/35717353/split-violin-plot-with-ggplot2}
}
\seealso{
\code{\link{geom_violin}}
}
\author{
jan-glx
}
