% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lighten_color.R
\name{lighten_color}
\alias{lighten_color}
\title{Lighten a color}
\usage{
lighten_color(color, extent = 100, preset = c(black = "transparent"))
}
\arguments{
\item{color}{a vector of color in any of the three available color types,
such as a color name, \code{#rrggbb}, and an integer. refer to
\code{\link{col2rgb}} in \code{\link{grDevices}} package.}

\item{extent}{the extent of lightening of color, where max value is 255.}

\item{preset}{previously selected conversion, which is given as a named
vector. The default, \code{c(black= "transparent")}, converts black to
transparent.}
}
\value{
a vector of lightened color in RGB format
}
\description{
Get a slightly brighter color
}
\examples{
lighten_color(c("green4", 8, 1, "black", "#6464FF", "blue"))

}
\seealso{
\code{\link{rgb}}
}
\author{
Mun-Gwan Hong, \email{mungwan@gmail.com}
}
