% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeToX_1.0.R
\name{mergeToX}
\alias{mergeToX}
\title{merge x and y, leaving x intact}
\usage{
mergeToX(
  x,
  y,
  by = intersect(names(x), names(y)),
  by.x = by,
  by.y = by,
  suffix.y = ".y"
)
}
\arguments{
\item{x, y}{data frames to be merged, in which 'x' will not be changed at all}

\item{by}{names of the columns in which the values link two data frames. When 
row names is used to match two tables, the function \code{rownames} 
can be given here.}

\item{by.x, by.y}{linking column names in 'x' and 'y' if the names are different}

\item{suffix.y}{the suffix for the columns in 'y' that have same name as 'x' 
excluding 'by.y'}
}
\value{
A merged data frame
}
\description{
merge two data frames, \code{x} and \code{y}, like \code{merge}, but \code{x}
is left intact
}
\examples{
tg1 <- data.frame("id"= c("s1", "s2", "s3", "s5"), "height"= c(169, 187, 175, 162))
tg2 <- data.frame("id"= c("s1", "s2", "s4", "s5"), "weight"= c(76, 89, 60, 68))
mergeToX(tg1, tg2)
rownames(tg1)
mergeToX(tg1, tg2, by= rownames)

}
\seealso{
\code{\link{merge}}
}
\author{
Mun-Gwan Hong, \email{mungwan@gmail.com}
}
