% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/names_by.R
\name{names_by}
\alias{names_by}
\title{Get names of 'by' output}
\usage{
names_by(x)
}
\arguments{
\item{x}{an \code{by} object}
}
\value{
a character vector with names
}
\description{
To the usual names of \code{\link{by}}-class as a list, the \var{INDICES} are
added to make the names clearer. This uses a fairly large part of the code of
\code{\link{print.by}} with some modification.
}
\seealso{
\code{\link{by}}, \code{\link{names}}
}
\author{
Mun-Gwan Hong, \email{mungwan@gmail.com}
}
