% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overwrite_par.R
\name{overwrite_par}
\alias{overwrite_par}
\title{Overwrite parameters}
\usage{
overwrite_par(prev_par, new_par, excl = c())
}
\arguments{
\item{prev_par}{a \code{list} of parameters on which \var{new_par} will 
overwrite.}

\item{new_par}{a \code{list} of new parameters}

\item{excl}{the parameter name that will be ignored from the \var{new_par}}
}
\value{
a \code{list} of parameters after overwriting
}
\description{
Over-write parameter list with new list
}
\examples{
prev_par <- list(x = 1:10, y= 10:1, col= "blue", cex= 0.5)
new_par <- list(col= "green", main= "Hello")
overwrite_par(prev_par, new_par)

}
\author{
Mun-Gwan Hong, \email{mungwan@gmail.com}
}
