% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panel.cor.R
\name{panel.cor}
\alias{panel.cor}
\title{show correlation value}
\usage{
panel.cor(x, y, digits = 2, cex.cor, method = "spearman", ...)
}
\arguments{
\item{x, y, method}{same as those for \code{\link{cor}}}

\item{digits}{how many digit of correlation values should be displayed.}

\item{cex.cor}{the magnification factor for correlation values}

\item{...}{not used}
}
\description{
This show correlation value whose color and size of font reflect how high the
value are. This is designed to use with \code{\link{pairs}} as \cr 
\code{pairs(x, upper.panel= panel.cor)}
}
\seealso{
\code{\link{pairs}}
}
\author{
Mun-Gwan Hong, \email{mun-gwan.hong@scilifelab.se}
}
