% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_lm.R
\name{plot_lm}
\alias{plot_lm}
\title{Plot - linear model}
\usage{
plot_lm(formula, data, ..., subset, res.pos = "topleft", res.inset = 0.5,
  abline_par = list(col = "cadetblue"), lm_txt_par = list())
}
\arguments{
\item{formula, data, subset}{same as those of \code{\link{plot.formula}} and
\code{\link{lm}} generic functions}

\item{...}{arguments to be passed to \code{\link{plot}}}

\item{res.pos}{a text that indicates where the results of linear model is to
be shown. The possible choices are \code{topleft}(default), \code{top},
\code{topright}, \code{bottomleft}, \code{bottom}, \code{bottomright},
\code{left}, \code{center}, \code{right}}

\item{res.inset}{the gap between the text of result and boundary. The value
is the ratio of gap to character size}

\item{abline_par}{the parameters passed to \code{'abline'} for the trend line}

\item{lm_txt_par}{the parameters passed to \code{'text'} that display the
coefficient and P-value from the linear regression model}
}
\value{
Same as output of the function of \code{\link{lm}}.
}
\description{
A scatter plot with linear regression results. In addition to a ordinary
scatter plot by \code{plot.formula}, a trend line estimated by linear
regression is displayed. The slope and significance are presented at the spot
of \code{res.pos}. Please note that this function accepts only one
independent variable in formula.
}
\examples{
df <- data.frame(x= 1:20)
df$y <- 0.2 * df$x + rnorm(20)
plot_lm(y ~ x, df, subset= x != 15, col= rep("blue", 20))

plot_lm(y ~ x, df, subset= x != 15, 
        lm_txt_par= list(label= coef(summary(lm(y ~ x, df)))[2, 4]))
  
}
\seealso{
\code{\link{plot}}, \code{\link{lm}}
}
\author{
Mun-Gwan Hong, \email{mun-gwan.hong@scilifelab.se}
}
