% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qqplot_log10P.R
\name{qqplot_log10P}
\alias{qqplot_log10P}
\title{Q-Q plot of P-values}
\usage{
qqplot_log10P(x, main = "Q-Q Plot of P-values", alpha = NULL,
  adjusted_p = NULL, cex = 0.3, pch = 1, ..., hit_txt_par = list(),
  legend_par = list())
}
\arguments{
\item{x}{a vector, matrix, or data frame of P-values}

\item{main, cex, pch}{same as those for \code{'plot'} generic function (refer
\code{\link{par}})}

\item{alpha}{a P-value threshold. Any hit in the vector or the first column
given by \var{x} below this value will be highlighted by showing the ID of
the hit. Note that this threshold will be used after multiple testing
correction.}

\item{adjusted_p}{The P-values after multiple testing correction. This is
valid only when \var{alpha} was given for the vector or the first column of
\var{x}. If this \code{vector} is not given, the default is Bonferroni
corrected values.}

\item{...}{arguments to be passed to \code{\link{qqplot}}}

\item{hit_txt_par}{a list of parameter for displaying hits}

\item{legend_par}{a list of parameter for a legend which will be presented
only when multiple columns and the names of them are given in \var{x}.}
}
\value{
the vector of hit names
}
\description{
Quantile-Quantile plot of observed P values vs. theorical P values in uniform
distribution
}
\seealso{
\code{\link{qqplot}}
}
\author{
Mun-Gwan Hong, \email{mun-gwan.hong@scilifelab.se}
}
