% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Olink_read_excel.R
\name{read_Olink_excel}
\alias{read_Olink_excel}
\title{Read Olink Excel}
\usage{
read_Olink_excel(file)
}
\arguments{
\item{file}{a Excel file from the manager}
}
\value{
a list of two tibbles (or data frames). One is for NPX data and the
  other for binders
}
\description{
This parse an output file in Excel format from Olink NPX manager
}
\seealso{
\code{\link{read_excel}}
}
\author{
Mun-Gwan Hong <\email{mun-gwan.hong@scilifelab.se}>
}
