% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shuffle.R
\name{shuffle}
\alias{shuffle}
\title{Shuffle}
\usage{
shuffle(x, size, replace = FALSE, prob = NULL)
}
\arguments{
\item{x, size, replace, prob}{same as \code{\link{sample}}}
}
\value{
a vector
}
\description{
Same as \code{\link{sample}} but it avoids the problem occured by a variable
size of \var{x}
}
\details{
This function is almost identical to \code{\link{sample}}. But, if 'x' was 
given with a single value, this returns the given value instead of the random
sampling of 1:(the value) of the \code{sample}.
}
\seealso{
\code{\link{sample}}
}
\author{
Mun-Gwan Hong, \email{mun-gwan.hong@scilifelab.se}
}
