% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_log.R
\name{source_log}
\alias{source_log}
\title{Source & log}
\usage{
source_log(file, echo = F)
}
\arguments{
\item{file}{the name of R file to be run}

\item{echo}{same as 'echo' in \code{\link{source}}}
}
\description{
Execute a script while writing a log to keep the history. The log file is
created at "../log" as a default. If it is not available, the file can be
found where the R script \var{file} is. Automatically, execution date is
added into the log file name. Loaded packages in the script are listed in the
log.
}
\seealso{
\code{\link{source}}
}
\author{
Mun-Gwan Hong, \email{mun-gwan.hong@scilifelab.se}
}
