% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string_utils.R
\name{str_rev}
\alias{str_rev}
\title{Reverse string}
\usage{
str_rev(x)
}
\arguments{
\item{x}{a string vector}
}
\value{
a string vector after reversing the order of characters in each
  string
}
\description{
Reverse the string. A copy from an example of \code{\link{strsplit}}
}
\examples{
str_rev(c("Sample1", "Sample2"))

}
\author{
Mun-Gwan Hong, \email{mungwan@gmail.com}
}
