% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string_utils.R
\name{toupper_1st_char}
\alias{toupper_1st_char}
\title{Cap 1st Character}
\usage{
toupper_1st_char(x)
}
\arguments{
\item{x}{a string vector}
}
\value{
a string vector
}
\description{
Capitalize the first character only
}
\author{
Mun-Gwan Hong, \email{mungwan@gmail.com}
}
