% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unwhich.R
\name{unwhich}
\alias{unwhich}
\title{Unwhich}
\usage{
unwhich(x, length)
}
\arguments{
\item{x}{a vector of numeric indicies which elements will be marked as TRUE}

\item{length}{the length of output vector}
}
\value{
a logical vector
}
\description{
Opposite of \code{\link{which}}. Convert to numeric index into a logical
vector
}
\examples{
unwhich(c(5, 10), 10)

}
\author{
Mun-Gwan Hong, \email{mun-gwan.hong@scilifelab.se}
}
