% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use.R_template.R
\name{use.R_template}
\alias{use.R_template}
\title{Use .R template}
\usage{
use.R_template(
  file,
  author,
  title = sub("\\\\.R$", "", basename(file)),
  date = Sys.Date()
)
}
\arguments{
\item{file}{The name of the generating .R file. The ".R" extension will be
added if it is missing.}

\item{author}{The author of the script}

\item{title}{The title of the generating R script, which is displayed at the
top}

\item{date}{The date written as the creation date}
}
\description{
Create .R file using a template that includes header lines
}
\examples{
\dontrun{
use.R_template("process_data") 
}

}
\author{
Mun-Gwan Hong, \email{mungwan@gmail.com}
}
