% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0.Useful4me-package.R
\docType{package}
\name{Useful2me-package}
\alias{Useful2me-package}
\title{Useful2me : A package of useful functions to me}
\description{
This package includes functions that simplify works here and there
}
\author{
Mun-Gwan Hong <\email{mungwan@gmail.com}>
}
\keyword{package}
