% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0.Useful4me-package.R
\docType{package}
\name{Useful4me-package}
\alias{Useful4me-package}
\title{Useful4me : A package of useful functions for me}
\description{
This package includes the functions simplifying works here and there
}
\author{
Mun-Gwan Hong <\email{mun-gwan.hong@scilifelab.se}>
}
\keyword{package}
