% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use.R_template.R
\name{use.R_template}
\alias{use.R_template}
\title{Use .R template}
\usage{
use.R_template(file, title = sub("\\\\.R$", "", basename(file)),
  author = "Mun-Gwan", date = Sys.Date(), wddir = getwd())
}
\arguments{
\item{file}{The name of the generating .R file. The ".R" extension will be
added if it is missing.}

\item{title}{The title of the generating R script, which is displayed at the
top}

\item{author}{The author of the script}

\item{date}{The date written as the creation date}

\item{wddir}{The directory on which the R script should be run.}
}
\description{
Create .R file using a template that includes header lines
}
\examples{
\dontrun{
use.R_template("process_data") 
}

}
\author{
Mun-Gwan Hong, \email{mun-gwan.hong@scilifelab.se}
}
